refactor: 레이싱경주의 우승자를 구함
도메인의 함수를 이동 (Racing -> Cars)
