+++ src/main/java/study/step4/Racing.java
@@ -1,7 +1,6 @@
 package study.step4;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Racing {
 
@@ -35,13 +34,6 @@ public class Racing {
     }
 
     public List<Car> getWinners() {
-        int maxLocation = cars.stream()
-                .map(car -> car.getLocation())
-                .max(Integer::compare)
-                .get();
-
-        return cars.stream()
-                .filter(car -> car.getLocation() == maxLocation)
-                .collect(Collectors.toList());
+        return cars.getWinners();
     }
 }
