+++ src/test/java/study/step4/CarsTest.java
@@ -1,9 +1,12 @@
 package study.step4;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -32,5 +35,25 @@ class CarsTest {
         assertThat(cars.stream()
                 .map(car -> car.getLocationToString())
                 .collect(Collectors.joining(RacingConstant.COMMA))).isEqualTo(expected);
+    }
+
+
+    @DisplayName("레이싱게임의 우승자 확인")
+    @Test
+    void checkWinner() {
+        List<Car> carList = new ArrayList<>();
+        Car car1 = Car.of("hwang");
+        Car car2 = Car.of("lee");
+        Car car3 = Car.of("kim");
+        car1.move(5);
+        car2.move(4);
+        car3.move(10);
+        carList.add(car1);
+        carList.add(car2);
+        carList.add(car3);
+
+        Cars cars = Cars.of(carList);
+
+        assertThat(cars.getWinners().get(0).getCarName()).isEqualTo("kim");
     }
 }
