+++ src/test/java/study/step4/RacingTest.java
@@ -1,12 +1,8 @@
 package study.step4;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -29,24 +25,5 @@ public class RacingTest {
             racing.race();
         }
         assertThat(racing.isFinished()).isTrue();
-    }
-
-    @DisplayName("레이싱게임의 우승자 확인")
-    @Test
-    void checkWinner() {
-        List<Car> carList = new ArrayList<>();
-        Car car1 = Car.of("hwang");
-        Car car2 = Car.of("lee");
-        Car car3 = Car.of("kim");
-        car1.move(5);
-        car2.move(4);
-        car3.move(10);
-        carList.add(car1);
-        carList.add(car2);
-        carList.add(car3);
-
-        Racing racing = Racing.of(Cars.of(carList), 5, new RandomMoveBehavior());
-
-        assertThat(racing.getWinners().get(0).getCarName()).isEqualTo("kim");
     }
 }
